+++ components/renku_data_services/data_connectors/api.spec.yaml
@@ -70,226 +70,6 @@ paths:
           $ref: "#/components/responses/Error"
       tags:
         - data_connectors
-  /data_connectors/storage:
-    post:
-      summary: Create a new project storage
-      requestBody:
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: "#/components/schemas/ProjectStoragePost"
-      responses:
-        "201":
-          description: The data connector was created
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ProjectStorage"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
-  /data_connectors/storage/config:
-    get:
-      summary: Get the current configuration for project storage
-      responses:
-        "200":
-          description: The configuration data
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ProjectStorageConfig"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
-  /data_connectors/storage/allow:
-    get:
-      summary: List all projects in the storage allow list
-      parameters:
-        - in: query
-          description: query parameters
-          name: params
-          style: form
-          explode: true
-          schema:
-            $ref: "#/components/schemas/ProjectStorageAllowListQuery"
-      responses:
-        "200":
-          description: List of storage allow entries
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ProjectStorageAllowList"
-          headers:
-            page:
-              description: The index of the current page (starting at 1).
-              required: true
-              schema:
-                type: integer
-            per-page:
-              description: The number of items per page.
-              required: true
-              schema:
-                type: integer
-            total:
-              description: The total number of items.
-              required: true
-              schema:
-                type: integer
-            total-pages:
-              description: The total number of pages.
-              required: true
-              schema:
-                type: integer
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
-    post:
-      summary: Add a project to the storage allow list
-      requestBody:
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: "#/components/schemas/ProjectStorageAllowPost"
-      responses:
-        "201":
-          description: The project was added to the allow list
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ProjectStorageAllowPost"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
-  /data_connectors/storage/allow/{project_id}:
-    parameters:
-      - in: path
-        name: project_id
-        required: true
-        schema:
-          $ref: "#/components/schemas/Ulid"
-    get:
-      summary: Get the storage allow entry for a project
-      responses:
-        "200":
-          description: The project storage allow entry
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ProjectStorageAllow"
-        "404":
-          description: The project is not in the storage allow list
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ErrorResponse"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
-    patch:
-      summary: Change the maximum size for a project storage
-      parameters:
-        - $ref: "#/components/parameters/If-Match"
-      requestBody:
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: "#/components/schemas/ProjectStorageAllowPatch"
-      responses:
-        "200":
-          description: The patched project storage allow entry
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ProjectStorageAllow"
-        "404":
-          description: The project storage allow entry doesn't exist
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ErrorResponse"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
-    delete:
-      summary: Remove a project from the storage allow list
-      responses:
-        "204":
-          description: The project was removed from the allow list
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
-  /data_connectors/storage/{storage_id}:
-    parameters:
-      - in: path
-        name: storage_id
-        required: true
-        schema:
-          $ref: "#/components/schemas/Ulid"
-    get:
-      summary: Get a project storage for a project id.
-      responses:
-        "200":
-          description: The project storage information
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ProjectStorage"
-        "404":
-          description: The project storage does not exist
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ErrorResponse"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
-    patch:
-      summary: Change the size or mount path of a project storage
-      parameters:
-        - $ref: "#/components/parameters/If-Match"
-      requestBody:
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: "#/components/schemas/ProjectStoragePatch"
-      responses:
-        "200":
-          description: The patched project storage entry
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ProjectStorage"
-        "404":
-          description: The project storage does not exist
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ErrorResponse"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
-    delete:
-      summary: Delete a specific project storage
-      responses:
-        "204":
-          description: The project storage was deleted or did not exist in the first place
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - data_connectors
   /data_connectors/global:
     post:
       summary: Create a new data connector
@@ -702,26 +482,6 @@ paths:
           $ref: "#/components/responses/Error"
       tags:
         - data_connectors
-  /projects/{project_id}/storage:
-    parameters:
-      - in: path
-        name: project_id
-        required: true
-        schema:
-          $ref: "#/components/schemas/Ulid"
-    get:
-      summary: Get the project storage for a given project
-      responses:
-        "200":
-          description: The list of project storages (currently either one or empty).
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ProjectStorageList"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - projects
   /projects/{project_id}/data_connector_links:
     parameters:
       - in: path
