+++ components/renku_data_services/storage/apispec.py
@@ -1,13 +1,13 @@
 # generated by datamodel-codegen:
 #   filename:  api.spec.yaml
-#   timestamp: 2026-07-30T05:48:01+00:00
+#   timestamp: 2026-07-30T15:37:23+00:00
 
 from __future__ import annotations
 
 from enum import StrEnum
 from typing import Any
 
-from pydantic import Field, RootModel
+from pydantic import ConfigDict, Field, RootModel
 from renku_data_services.storage.apispec_base import BaseAPISpec
 
 
@@ -100,6 +100,29 @@ class ErrorResponse(BaseAPISpec):
     error: Error
 
 
+class ProjectStoragePatch(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    size: int | None = Field(None, description="The maximum size in GB")
+    mount_path: str | None = Field(None, description="The mount path for the storage")
+
+
+class ProjectStorageAllowPatch(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    max_size: int | None = Field(None, description="The maximum size in GB")
+
+
+class ProjectStorageConfig(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    enabled: bool
+    max_size: int = Field(..., description="The maximum size in GB")
+
+
 class StorageSchemaTestConnectionPostRequest(BaseAPISpec):
     configuration: dict[str, int | str | bool | dict[str, Any] | None]
     source_path: str = Field(
@@ -113,6 +136,14 @@ class StorageSchemaObscurePostRequest(BaseAPISpec):
     configuration: dict[str, int | str | bool | dict[str, Any] | None]
 
 
+class PaginationRequest(RootModel[Any]):
+    root: Any
+
+
+class ProjectStorageList(RootModel[Any]):
+    root: Any
+
+
 class RCloneEntry(BaseAPISpec):
     name: str = Field(..., description="Human readable name of the provider")
     description: str = Field(..., description="description of the provider")
@@ -120,6 +151,62 @@ class RCloneEntry(BaseAPISpec):
     options: list[RCloneOption] = Field(
         ..., description="Fields/properties used for this storage."
     )
+
+
+class ProjectStoragePost(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    namespace: Any
+    size: int
+    mount_path: str
+
+
+class ProjectStorage(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    id: Any
+    project_id: Any
+    size: int
+    mount_path: str
+    creation_date: Any
+    created_by: Any
+    updated_at: Any
+    etag: Any
+
+
+class ProjectStorageAllowPost(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    project_id: Any
+    max_size: int = Field(..., description="Maximum size in GB")
+
+
+class ProjectStorageAllow(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    project_id: Any
+    name: str
+    namespace: str
+    max_size: int = Field(..., description="Maximum size in GB")
+    etag: Any
+
+
+class ProjectStorageAllowList(RootModel[list[ProjectStorageAllow]]):
+    root: list[ProjectStorageAllow] = Field(
+        ..., description="A list of project storage allow entries.\n"
+    )
+
+
+class ProjectStorageAllowListQuery(BaseAPISpec):
+    project_name: str = Field("", description="Filter by project name (partial match).")
+
+
+class DataConnectorsStorageAllowGetParametersQuery(BaseAPISpec):
+    params: ProjectStorageAllowListQuery | None = None
 
 
 class RCloneSchema(RootModel[list[RCloneEntry]]):
