+++ components/renku_data_services/storage/models.py
@@ -1,6 +1,11 @@
-"""Models for cloud storage."""
+"""Models for storage."""
+
+from __future__ import annotations
 
 from collections.abc import Generator, MutableMapping
+from dataclasses import dataclass
+from datetime import datetime
+from pathlib import PurePosixPath
 from typing import Any
 from urllib.parse import ParseResult, urlparse
 
@@ -8,7 +13,12 @@ from pydantic import BaseModel, Field, PrivateAttr, model_serializer, model_vali
 from ulid import ULID
 
 from renku_data_services import errors
+from renku_data_services.base_models.bytesize import ByteSize
+from renku_data_services.base_models.core import (
+    ProjectPath,
+)
 from renku_data_services.storage.rclone import RCloneValidator
+from renku_data_services.utils.etag import compute_etag_from_fields
 
 
 class RCloneConfig(BaseModel, MutableMapping):
@@ -19,7 +29,7 @@ class RCloneConfig(BaseModel, MutableMapping):
     _validator: RCloneValidator = PrivateAttr(default=RCloneValidator())
 
     @model_validator(mode="after")
-    def check_rclone_schema(self) -> "RCloneConfig":
+    def check_rclone_schema(self) -> RCloneConfig:
         """Validate that the reclone config is valid."""
         self._validator.validate(self.config)
         return self
@@ -71,7 +81,7 @@ class UnsavedCloudStorage(BaseModel):
     """Path inside the target repository to mount/clone data to."""
 
     @classmethod
-    def from_dict(cls, data: dict) -> "UnsavedCloudStorage":
+    def from_dict(cls, data: dict) -> UnsavedCloudStorage:
         """Create the model from a plain dictionary."""
 
         if "project_id" not in data:
@@ -101,7 +111,7 @@ class UnsavedCloudStorage(BaseModel):
     @classmethod
     def from_url(
         cls, storage_url: str, name: str, readonly: bool, project_id: str, target_path: str
-    ) -> "UnsavedCloudStorage":
+    ) -> UnsavedCloudStorage:
         """Get Cloud Storage/rclone config from a storage URL.
 
         Example:
@@ -132,7 +142,7 @@ class UnsavedCloudStorage(BaseModel):
     @classmethod
     def from_s3_url(
         cls, storage_url: ParseResult, project_id: str, name: str, readonly: bool, target_path: str
-    ) -> "UnsavedCloudStorage":
+    ) -> UnsavedCloudStorage:
         """Get Cloud storage from an S3 URL.
 
         Example:
@@ -176,7 +186,7 @@ class UnsavedCloudStorage(BaseModel):
     @classmethod
     def from_azure_url(
         cls, storage_url: ParseResult, project_id: str, name: str, readonly: bool, target_path: str
-    ) -> "UnsavedCloudStorage":
+    ) -> UnsavedCloudStorage:
         """Get Cloud storage from an Azure URL.
 
         Example:
@@ -212,7 +222,7 @@ class UnsavedCloudStorage(BaseModel):
     @classmethod
     def _from_ambiguous_url(
         cls, storage_url: ParseResult, project_id: str, name: str, readonly: bool, target_path: str
-    ) -> "UnsavedCloudStorage":
+    ) -> UnsavedCloudStorage:
         """Get cloud storage from an ambiguous storage url."""
         if storage_url.hostname is None:
             raise errors.ValidationError(message="Storage URL must contain a host")
@@ -228,3 +238,109 @@ class CloudStorage(UnsavedCloudStorage):
     """Cloudstorage saved in the database."""
 
     storage_id: ULID = Field()
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class UnsavedProjectStorage:
+    """Project storage definition."""
+
+    namespace_path: ProjectPath
+    size: ByteSize
+    mount_path: PurePosixPath
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class ProjectStoragePatch:
+    """Model for changes requested on a project storage."""
+
+    size: ByteSize | None
+    mount_path: PurePosixPath | None
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class ProjectStorage:
+    """Stored project storage information."""
+
+    id: ULID
+    project_id: ULID
+    storage_class: str
+    size: ByteSize
+    mount_path: PurePosixPath
+    created_by: str
+    creation_date: datetime
+    updated_at: datetime
+
+    @property
+    def etag(self) -> str:
+        """Entity tag value for this project storage object."""
+        return compute_etag_from_fields(
+            self.updated_at, self.project_id, self.storage_class, self.size.to_bytes(), self.mount_path.as_posix()
+        )
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class DeletedProjectStorage:
+    """A project storage that has been deleted."""
+
+    project_id: ULID
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class ProjectStorageAllow:
+    """Allowed project storage with max size."""
+
+    project_id: ULID
+    max_size: ByteSize
+    updated_at: datetime
+
+    @property
+    def etag(self) -> str:
+        """Entity tag value for this project storage allow object."""
+        return compute_etag_from_fields(self.updated_at, self.project_id, self.max_size.to_bytes())
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class ProjectStorageAllowDetail:
+    """Allowed project storage with max size."""
+
+    project_id: ULID
+    max_size: ByteSize
+    name: str
+    namespace_path: ProjectPath
+    updated_at: datetime
+
+    @classmethod
+    def create(
+        cls,
+        project_id: ULID,
+        max_size: ByteSize,
+        name: str,
+        namespace_slug: str,
+        project_slug: str,
+        updated_at: datetime,
+    ) -> ProjectStorageAllowDetail:
+        """Create an instance with the project path given as two strings."""
+        np = ProjectPath.from_strings(namespace_slug, project_slug)
+        return ProjectStorageAllowDetail(
+            project_id=project_id, max_size=max_size, name=name, namespace_path=np, updated_at=updated_at
+        )
+
+    @property
+    def etag(self) -> str:
+        """Entity tag value for this project storage allow object."""
+        return compute_etag_from_fields(self.updated_at, self.project_id, self.max_size.to_bytes())
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class ProjectStorageAllowPatch:
+    """Model for changes requested on a project storage allow entry."""
+
+    max_size: ByteSize | None
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class ProjectStorageAllowUpdate:
+    """Return data when updating an allow entry."""
+
+    old: ProjectStorageAllowDetail
+    new: ProjectStorageAllowDetail
